:recycle:refactor : 변수명(set/get X)정리
